refactor: SubwayAcceptanceTest 추상 클래스 추가
