+++ src/test/java/nextstep/subway/SubwayAcceptanceTest.java
@@ -1,0 +1,21 @@
+package nextstep.subway;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+
+import io.restassured.RestAssured;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public abstract class SubwayAcceptanceTest {
+
+	@LocalServerPort
+	int port;
+
+	@BeforeEach
+	void setUp() {
+		if (RestAssured.port == RestAssured.UNDEFINED_PORT) {
+			RestAssured.port = port;
+		}
+	}
+}
